fix: refactoring current pages and components
