+++ src/screens/Categories/categories.routes.tsx
@@ -1,7 +1,7 @@
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import React from "react";
 import { Categories } from ".";
-import { CategoryDetails } from "./Details";
+import { CategoryDetails } from "../CategoryDetails";
 const { Screen, Navigator } = createNativeStackNavigator();
 
 export function CategoriesStackRoutes() {
