+++ src/screens/DrinkDetails/index.tsx
@@ -1,0 +1,3 @@
+export function DrinkDetails() {
+  return <h1>Teste</h1>;
+}
